+++ Library/Adjustment.cs
@@ -47,7 +47,7 @@ namespace Recurly
         public DateTime StartDate { get; protected set; }
         public DateTime? EndDate { get; protected set; }
 
-        public DateTime CreatedAt { get ; protected set; }
+        public DateTime? CreatedAt { get ; protected set; }
 
         private const string UrlPrefix = "/accounts/";
         private const string UrlPostfix = "/adjustments/";
@@ -208,7 +208,10 @@ namespace Recurly
                         break;
 
                     case "created_at":
-                        CreatedAt = reader.ReadElementContentAsDateTime();
+                      //  CreatedAt = reader.ReadElementContentAsDateTime();
+                        DateTime createdAt;
+                        if (DateTime.TryParse(reader.ReadElementContentAsString(), out createdAt))
+                            CreatedAt = createdAt;
                         break;
 
                     case "state":